# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
...
        "/user/classes/enrolled/{userName}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "수강하고 있는 강의목록 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "유저 아이디",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "강의목록(배열)",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.Class"
                            }
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        },
...
        "/user/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "로그인",
                "parameters": [
                    {
                        "description": "로그인 정보",
                        "name": "JSON",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.Login"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "USER_NOT_EXIST",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "회원가입",
                "parameters": [
                    {
                        "description": "유저정보",
                        "name": "JSON",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        },
        "/users/{username}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "유저 로그인 아이디 중복 확인",
                "parameters": [
                    {
                        "type": "string",
                        "description": "확인 하려는  로그인 아이디",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "USER_NOT_EXIST",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.Class": {
            "type": "object",
            "properties": {
                "class_code": {
                    "type": "string",
                    "example": "형식언어"
                },
                "class_name": {
                    "type": "string",
                    "example": "CSE-1234"
                }
            }
        },
        "dto.Login": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "dto.User": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "donggukmail@dgu.ac.kr"
                },
                "nickname": {
                    "type": "string",
                    "example": "my nickname"
                },
                "student_code": {
                    "type": "integer",
                    "example": 2015123456
                },
                "username": {
                    "type": "string",
                    "example": "pigeon1234"
                }
            }
        }
    }
}